requirements.txt
numpy==1.23.4
python-magic-bin==0.4.14
Pillow==9.3.0
pillow-heif==0.7.1
